+++ docs/changelog.rst
@@ -19,7 +19,7 @@ Changelog
 * new: depend on Autobahn v21.11.1
 * new: expand WAP web service (#1878)
 * fix: various adjustments and fixes after integration of FX code base
-* new: open-source code for "Crossbar.io FX" (~26k LOC), incl. router-to-router links
+* new: open-source code for "Crossbar.io" (~26k LOC), incl. router-to-router links
 * new: changed license from AGPLv3 to [EUPLv1.2](https://eupl.eu/1.2/en) (under IP ownership of Crossbar.io Technologies GmbH)
 
 21.3.1